abstract chat to a component
